Moved AuthService to shared/services
