+++ angular-client/src/app/modules/authentification/authentication.component.ts
@@ -1,5 +1,5 @@
 import { Component } from "@angular/core";
-import { AuthService } from "./auth.service";
+import { AuthService } from "../../shared/services/auth.service";
 
 @Component({
   selector: 'app-authentication',
@@ -11,6 +11,6 @@ export class AuthenticationComponent {
   constructor(private authService: AuthService){}
 
     isLoggedIn(){
-      return this.authService.isLoggedIn();
+      return AuthService.isLoggedIn();
     }
   }
