+++ angular-client/src/app/modules/authentification/can-activate.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { CanActivate } from '@angular/router';
-import { AuthService } from './auth.service';
+import { AuthService } from '../../shared/services/auth.service';
 
 @Injectable()
 export class CanActivateViaAuthGuard implements CanActivate {
@@ -8,6 +8,6 @@ export class CanActivateViaAuthGuard implements CanActivate {
   constructor(private authService: AuthService) {}
 
   canActivate() {
-    return this.authService.isLoggedIn();
+    return AuthService.isLoggedIn();
   }
 }
